ai-review-engine_updated/requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
scrapy==2.11.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil==2.8.2

# NLP and Text Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
vaderSentiment==3.3.2
wordcloud==1.9.3

# Machine Learning & Deep Learning
scikit-learn>=1.3.0
transformers>=4.36.0
torch>=2.1.0
tensorflow>=2.15.0
pytorch-lightning>=2.1.0
sentence-transformers==2.2.2

# Advanced AI Models
optuna==3.5.0  # AutoML
huggingface-hub==0.20.1
tokenizers==0.15.0
datasets==2.16.1
accelerate==0.26.1

# Aspect-Based Sentiment Analysis
pyabsa==2.3.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1  # Database migrations
pymongo==4.6.1

# Web Framework
# Core Dependencies (Required)
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.15.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Web Scraping
beautifulsoup4>=4.12.0
selenium>=4.10.0
requests>=2.31.0

# Data Processing
scipy>=1.10.0
openpyxl>=3.1.0  # For Excel file handling

# Caching (Optional)
redis>=4.6.0

# Additional utilities
python-dotenv>=1.0.0
pyyaml>=6.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# API Framework
fastapi==0.105.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Message Queue & Caching
celery==5.3.4
redis==5.0.1
flower==2.0.1  # Celery monitoring

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
colorama==0.4.6
loguru==0.7.2  # Advanced logging

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Deployment
docker>=4.0.0
gunicorn==21.2.0

# Additional Advanced Features
pillow==10.2.0  # Image processing
opencv-python==4.9.0.80  # Computer vision
joblib==1.3.2  # Model persistence
wandb==0.16.2  # ML experiment tracking
mlflow==2.9.2  # Model versioning
